Update the parallel merge mesh stuff to use new tag_shared_verts function.
git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@4902 6162379e-bd28-0410-9a7d-b7f4dcfcad3c

+++ src/parallel/moab/ParallelMergeMesh.hpp
@@ -26,18 +26,17 @@ private:
   double myEps;
 
   static void temp_tuple_swap_real(tuple_list *tup, 
-				    unsigned a, 
-				   unsigned b);
+                                   unsigned long a, unsigned long b);
 
   static void temp_tuple_sort_real(tuple_list *tup,
 				   double eps2=0);
 
   static void temp_perform_sort_real(tuple_list *tup, 
-				     int left,
-				     int right,
-				     double eps2);
+                                     unsigned long left, 
+                                     unsigned long right,
+                                     double eps2);
 
-  static bool greaterThan(tuple_list *tup, int vrI,int vrJ, double eps2);
+  static bool greaterThan(tuple_list *tup, unsigned long vrI, unsigned long vrJ, double eps2);
 };
 
 } // namespace moab
